Removed global event var -- yield now returns an event object
Before, a __builtins__ hack was used to make the event object
available in robot.py. The new API (version 4) allows yield to
return a value -- this is what generator.send(val) does -- which
in this case is the event wrapped in an EventInfo object (to
allow for multiple events to be returned).