Use the __builtin__ module for modifying builtins
This is the documented way of doing it.  It makes much more sense.

However, it looks like it will be better to hook into the module
importing code in the future and scan through the module to find
decorated coroutines that way.  This might have the potential to
completely skip the hacky globals approach.