Shuffle files around a bit more.  'hardware' code now in hw/
I realised that the way I'd reorganised it just now meant that some
of the python modules that want to use SRIC would be imported in
places that they shouldn't be.

The things that interact with the hardware (e.g. motor.py, jointio.py,
etc.) now live in the "hw" directory.