Adjust vision object to work with new pyenv:
As it stands:
 * VisEvent represents an event. No suprise.
 * VisProc represents the hueblobs process
 * VisObj is the users "vision" handle on the vision system.

Whenever VisObj gets evaluated, we check to see if a request is in
flight - if it isn't, we start a new one. Repeated calls are made to
the VisProc polling routine to see whether or not our request has
come back, returning a VisEvent when it does.