Remove the shutdown stuff from sricd.py
This code is only ever run when there's a error in the user code.
sricd is required to scroll the log. The log is quite useful when code
fails ;)

sricd isn't cleanly shutdown when the stick is pulled anyway as everything
just get killed so we don't lose anything by doing this.