Add run.py to orchestrate the unwrapping of a fastwrapped robot.zip
This new run.py, found in /fastwrap,

This commit also moves fastwrap-extract into /fastwrap, as it's
required for executing

I've run a couple of tests.  These are the time between putting the
USB key in, and something coming up on the screen:

 1) With a non-fast-wrapped robot.zip: 19 seconds.
 2) With fast-wrapped

(Note that the above were both done on a 'warm' power board --
i.e. I'd already been through some interations of running pyenv, so
things like dbus and things will have been in cache already.)

So fastwrap saves 6 seconds off the start-up time, which seems worth
it to me.