Run user code as program from user/
Importing their code sucked for many reasons.  The main one being that
their code wasn't particularly normal-program-like.

This patch changes pyenv to look for the user's code in user/ rather
than an inner robot.zip.  This is beneficial because the contents of
the stuff that comes out of the IDE is less obfuscated, and the
robot.py that is inside user/ can just be executed without any fudging
around.