Add a class for managing threadlocal sric contexts
Instantiate one libsric/pysric context per thread.  When a thread that
hasn't yet got a context asks for one, create it.

This object (instance of SricCtxMan) will be passed into the
constructor of all the classes representing various bits of hardware.
When they require a sric context, they'll request it from this
manager.  This means that, for example, motor objects can be passed
between threads...