Wrap SricDevice objects up so they can be passed between threads
Move SricCtxMan out of robot.py and into into tssric.py, and add a new
class: TSSricDevice (Thread-Safe Sric Device).

TSSricDevice just passes attribute access through to the threadlocal
SricDevice that it wraps for that thread.  So, for example, an object
representing a motor controller will hold a TSSricDevice and use it as
it would a SricDevice.