Change the form of Poll.eval
Previously, when a poll evaluated to True, all the polls that were
being polled would be asked to populate an event object.  In the new
API, when a poll evaluates to True, as well as returning that the
event happened, it will also return information about the event to be
transferred to the user.