Convert And and Or to the new event-reporting situation
Also, expose them through the sr module.

When either of the And or Or operators do not detect an event, they
return None.  This has advantages and disadvantages:

For this example:

	r = wait_for( And( poll0, poll1 ) )

Disadvantage: When the And/Or returns False, one can't use r[0][0], as
it won't normally exist.

Advantage: One can check if the whole And() group returned an event
using r[0] == None, which I think is a much more useful thing.