Add a destructor to the Vision class
This adds the possibility of re-creating the vision object in the
Robot class to allow a webcam to be re-enumerated.

A team have expressed an interest in switching between two webcams at
runtime (by switching the power to them). An easy way I can see of
doing this is to delete the vision member of the Robot object and
re-creating it with Robot._init_vision().

Ideally they would be able to plug multiple webcams in at the same time
and select which one to use when calling Robot.see(), however this is a
quick and simple fix to allow rudimentary support for multiple webcams.

I've tested this new functionality with the following robot.py code:

-----
from sr import *
from time import sleep
R = Robot(wait_start=False)

while(1):
    for i in range(10):
        markers = R.see()
        print "I can see", len(markers), "markers"
        sleep(0.5)
    print "Switch webcams now"
    del R.vision
    sleep(5)
    R._init_vision("/dev/video0")
-----

Obviously this exploits some inner workings of pyenv that aren't publicly
documentated and are not guaranteed to stay the same over time but I don't
think it's too bad. It's up to the team in question if they're prepared
to accept this unstableness until a more comprehensive solution exists.