Load a precalculated version of ld.so.cache on startup
ctypes uses ldconfig to find libraries.  ldconfig caches its findings
about installed libraries in ld.so.cache.  We haven't shipped an
ld.so.cache on the BeagleBoard image, which means that from cold
ldconfig takes ages to run (it's like seconds...).  It would therefore
be inconvenient to run ldconfig from cold on every startup.

So this patch adds an ld.so.cache that I just generated on a
BeagleBoard to robot.zip.  It modifies run.py to install this cache
file in the right place upon startup.