Add a TSSricDevice wrapper providing a device-specific lock
Some things we would do with the boards involve multiple SRIC
transactions, such as setting a single output pin, or toggling an LED
on the power board.  These transactions were not protected by a lock,
so incorrect things would occur if two threads both attempted to
perform one of these multi-transaction operations at the same time.

This patch adds a wrapper class around TSSricDevice, which is the
class that all the Python libraries use to access devices at the
moment.  This wrapper class, called LockableDev, essentially adds one
thing to TSSricDevice: a lock.  The lock *must* be acquired before any
actions are performed on the LockableDev.

Python threading.Lock objects can be used with Python's 'with'
statement, which makes using this lock extremely straightforward.  For
example, given a LockableDev instance d:

with d.lock:
   d.txrx( ... )
   ... munge some things ...
   d.txrx( ... )

LockableDev also wraps the TSSricDevice's txrx function to check that
it is being called whilst the lock is held.  This will assist mainly
in development, catching cases where someone has attempted to do
something with the LockableDev without first acquiring the lock
first.  This performs no checking that the thread that holds the lock
is the thread that's attempting to do things.  So, this will only
assist in testing and will catch all fail in single-threaded cases,
which IMO makes it quite valuable...

Issue first described in trac #1096.