Annihilate globals passed from sr.loggrok to the user's context
The user's code is run using execfile().  This is so that we can
achieve the standard output newline mangling etc.  One effect of this
was that __file__ was set to the path of "loggrok.py".  Not only is
not what we want because it's confusing, but also it has highly
confusing results when one is trying to import files and things.  Had
a team tried to call a file "power.py" and import it, they would have
actually ended up with the "power.py" from the same directory as
loggrok.py.

This patch clears the globals passed to the user's code, and sets
__file__ to be the thing that they expect.  This clearing means that
importing "power.py" in a user's code does what one would expect:
looking for "power.py" in the same directory as the user's code!