Turn optimisation off in our Python environment
When Python's 'optimisation' is on, assert statements do not work.
This is somewhat irritating.  This is because in our OE build,
python was patched to default having Python on.  There's no Python- or
cli-based way of turning it off again.

So we use the python C API to make our own python interpreter.  It's
very short, as it just changes the value of Py_OptimizeFlag to 0 then
calls Py_Main...

This patch adds the wrapper from git://srobo.org/~rspanton/pynoopt.git
revision 7aceb096, and puts it in bin/python.  It also shuffles our
bin directory to the front of PATH so that we actually use it.

This results in assert statements working :-)

See the extended description of this in Trac #1110