Add support for motor controller enumeration
Use pyudev to find the motor controllers.  Store them in a dictionary,
where they can be accessed either with an integer index, or their
serial number.  This allows teams to simply use an integer, and then
later when they discover that this is annoying when they remove a
motor controller, the serial number.

Change-Id: I8e9e03c54a2e05482132897e4cfcf01c71ac9275