Work around a bug in the MCv4B:1 firmware
The version 1 firmware has a bug that means setting the speed/direction
of a channel sets the direction on that channel and the speed on the
other channel. This works around that problem by keeping track of the
state of the motor board and always setting both channels. The direction
is set on the correct channel (with the stored speed), and the speed set
on the other channel (with the stored direction).

Change-Id: Id47db025ec6973e462bb978fa06d30d85a050041