Fix enumeration of multiple USB devices
When multiple motor or ruggeduino boards were present, only one of
each type would be accessible using numeric keys -- e.g. if two motor
controllers were present, then only Robot.motors[0] would exist, and
not Robot.motors[1].

This was because a counter was not incremented when adding those items
to the dicts.

Change-Id: Ie4f19de21ac03767c49e030d796192e8171888e6