Optionally remove gunk files from user code
Sometimes it may be useful to remove temporary editor files from the
user code.  Under the principle of least surprise, we won't be
deleting files from the user's code by default.  So by default this
removal is turned off, and can be turned on with the '-r' switch.

Additionally, remove 'swp' files as well as nano save files.

Change-Id: I420e1eefc307d85ad8bfe34c218005aeb1f170b8