Retry reading mcv4 fw version in case of failure
On a cold boot, the motor rail is not initially on.  It is only
possible to communicate with the motor controllers when the motor rail
is powered-up.  There is a non-zero period of time between the motor
rail being turned on and the point when we can communicate with the
motor controller.

We were trying to read the firmware version out of the motor
controller too soon, before it had finished powering-up.  This patch
makes it retry the process of reading the firmware version up to 10
times before giving up.

This means that on cold-boot, the message "Warning: Motor board is not
running the expected firmware" will no longer appear unless the motor
board is actually running the wrong firmware.

Change-Id: I7c2eb050d97166cdb40f6fcff309d1ee0c4945e2