Dump detected USB devices to log on start-up
Since both the motor controllers and jointio replacement are now USB,
they don't appear in the SRIC device listing on start-up.  This patch
adds listing of the detected USB devices.  It also adds a "serialnum"
property to the Motor and Ruggeduino objects so that their serial
numbers can be displayed in this listing.

Change-Id: I23a310d644d5997d390652d0325e4e800bcb96ad