Send all log messages through the logging module
There are times when it is useful to be able to increase the verbosity
of the log messages from pyenv, as well as times when it is useful to
reduce that verbosity.  Python's logging module allows one to set
filters and handlers for different levels of logs, so we should use
that to allow one to configure those things.

Since we don't want to enforce our teams to have to configure the
logging module in order to write some code, the Robot() constructors
now configure the logger to send messages with the INFO level and
above to stdout by default.

If and when the user wishes to configure the logging otherwise, they
can set the "config_logging" argument passed to Robot() (or
Robot.setup()) to be False.  To see any log messages, they will then
have to configure the logging module themselves.

Change-Id: Id8f623bfb1d4e622284d20c673db04232acee041