Kill motors when the USB stick is removed
The destructor of the Motor object which sets the target power to 0
does not seem to be being called (it is not guaranteed that it will be
called anyway). Instead, I have attached a SIGTERM handler to run.py
which calls self.user.end_match() which, in turn, stops the motors by
killing all power to the motor rail.

Change-Id: I718bc5ff7276dd7fc10161f69a36a9c22d571155