Add support for updating motor board firmware
Add support to FwUpdater to check and update motor board firmwares.

This adapts sr.motor.Motor's existing ability to check the firmware
version of the device.  If it's not correct, then the InvalidFirmware
exception is raised, and the firmware is subsequently updated.

This checks that the bootloader has been entered after sending the
bootloader entry command to the motor board.

Change-Id: I1c49dd3dffb383a647017fa875b62d82d346f048