Require the user to press buttons before fw update
If a motor board firmware update is interrupted halfway through, then
the board will not work on the next power-up.  It'd require more user
intervention to unbrick it.

To try to avoid the effects of the user unplugging stuff due to
something surprisingly different coming up on the screen, this patch
adds a prompt that the user has to read before the firmware update
will proceed.  It ensures that the user has read the prompt by asking
them to press a particular set of buttons.  Once they do that, the
firmware goes ahead as before.

Since user input is required for the prompt, sr-input has to be
started slightly earlier.  If sricd is restarted, then sr-input also
has to be restarted.

Let's not go distributing some instructions that just tell the user
which buttons to press.  The prompt is self-documenting.  Other
documentation will just reduce the effectiveness of the prompt.

Change-Id: I40c886f976bf811797a8ac75e8e64f474f2aac61