Disable power board firmware updates
Some power boards appear to be updating, despite having the right
firmware version in them.  Rather than dig into this problem now,
disabling the power board updates is the fastest way to fix this.

Since all power boards will actually have the right firmware in them
by now (it was Feb 2012 that the last update went out), this is fine.

Change-Id: I6d805861f7e75656d027adc4a3f2eec278eb5c84